+++ relay-config/src/config.rs
@@ -2425,6 +2425,7 @@ impl Config {
         Some(build_redis_configs(
             redis_configs,
             self.cpu_concurrency() as u32,
+            self.pool_concurrency() as u32,
         ))
     }