+++ relay-server/src/services/processor/span/processing.rs
@@ -232,8 +232,16 @@ pub async fn process(
         };
         new_item.set_payload(ContentType::Json, payload);
         new_item.set_metrics_extracted(item.metrics_extracted());
-        new_item
-            .set_ingest_span_in_eap(project_info.config.features.has(Feature::IngestSpansInEap));
+        new_item.set_ingest_span_in_eap(
+            project_info
+                .config
+                .features
+                .has(Feature::IngestSpansInEapForOrganization)
+                || project_info
+                    .config
+                    .features
+                    .has(Feature::IngestSpansInEapForProject),
+        );
 
         *item = new_item;
 
@@ -299,7 +307,14 @@ pub fn extract_from_event(
         .envelope()
         .dsc()
         .and_then(|ctx| ctx.sample_rate);
-    let ingest_in_eap = project_info.config.features.has(Feature::IngestSpansInEap);
+    let ingest_in_eap = project_info
+        .config
+        .features
+        .has(Feature::IngestSpansInEapForOrganization)
+        || project_info
+            .config
+            .features
+            .has(Feature::IngestSpansInEapForProject);
 
     let mut add_span = |mut span: Span| {
         add_sample_rate(
@@ -1135,7 +1150,7 @@ mod tests {
                 }
             }"#,
         )
-        .unwrap();
+            .unwrap();
         populate_ua_fields(
             span.value_mut().as_mut().unwrap(),
             None,
@@ -1153,7 +1168,7 @@ mod tests {
                 }
             }"#,
         )
-        .unwrap();
+            .unwrap();
         populate_ua_fields(
             span.value_mut().as_mut().unwrap(),
             None,
@@ -1175,7 +1190,7 @@ mod tests {
                 }
             }"#,
         )
-        .unwrap();
+            .unwrap();
         populate_ua_fields(
             span.value_mut().as_mut().unwrap(),
             Some("Mozilla/4.0 (compatible; MSIE 6.0; Windows NT 5.1; ONS Internet Explorer 6.1; .NET CLR 1.1.4322)"),
@@ -1212,7 +1227,7 @@ mod tests {
                 }
             }"#,
         )
-        .unwrap();
+            .unwrap();
         populate_ua_fields(
             span.value_mut().as_mut().unwrap(),
             None,
@@ -1417,7 +1432,7 @@ mod tests {
               "segment_id": "88457c3c28f4c0c6"
         }"#,
         )
-        .unwrap();
+            .unwrap();
 
         normalize(&mut span, normalize_config()).unwrap();