+++ CHANGELOG.md
@@ -4,6 +4,22 @@
 
 **Features**:
 
+- Add experimental playstation endpoint. ([#4555](https://github.com/getsentry/relay/pull/4555))
+- Add naver.me / Yeti spider on the crawler filter list. ([#4602](https://github.com/getsentry/relay/pull/4602))
+
+**Bug Fixes**:
+
+- Separates profiles into backend and ui profiles. ([#4595](https://github.com/getsentry/relay/pull/4595))
+
+**Internal**:
+
+- Add ui chunk profiling data category. ([#4593](https://github.com/getsentry/relay/pull/4593))
+- Switch global rate limiter to a service. ([#4581](https://github.com/getsentry/relay/pull/4581))
+
+## 25.3.0
+
+**Features**:
+
 - Tag images with release version. ([#4532](https://github.com/getsentry/relay/pull/4532))
 - Switch default envelope compression from gzip to zstd. ([#4531](https://github.com/getsentry/relay/pull/4531))
 - Update release to include an aarch64 binary. ([#4514](https://github.com/getsentry/relay/pull/4514))
@@ -12,10 +28,13 @@
 - Update Apple device model classes ([#4529](https://github.com/getsentry/relay/pull/4529))
 - Remove separate quota and rate limit counting for replay videos ([#4554](https://github.com/getsentry/relay/pull/4554))
 - Deprecate ReplayVideo data category ([#4560](https://github.com/getsentry/relay/pull/4560))
+- Improve localhost detection by checking contained request headers in the error. ([#4580](https://github.com/getsentry/relay/pull/4580))
 
 **Bug Fixes**:
 
 - Prevent partial trims in indexed and queryable span data. ([#4557](https://github.com/getsentry/relay/pull/4557))
+- Emit filtered/sampled outcomes for spans attached to a dropped transaction. ([#4569](https://github.com/getsentry/relay/pull/4569))
+- Fix missing geo information for user when IP scrubbing was enabled. ([#4586](https://github.com/getsentry/relay/pull/4586))
 
 **Internal**: